+++ apps/meteor/client/views/account/sidebarItems.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { hasPermission, hasAtLeastOnePermission } from '../../../app/authorization/client';
 import { settings } from '../../../app/settings/client';
-import { defaultFeaturesPreview } from '../../hooks/useFeaturePreview';
+import { defaultFeaturesPreview } from '../../hooks/useFeaturePreviewList';
 import { createSidebarItems } from '../../lib/createSidebarItems';
 import AccountFeaturePreviewBadge from './featurePreview/AccountFeaturePreviewBadge';
